+++ src/resourceManager.ts
@@ -2,38 +2,42 @@ import { v5 as uuidv5 } from 'uuid';
 import { getVerifiedAuthentication, } from './storePrivateKeys.js';
 import { CheqdNetwork } from '@cheqd/sdk';
 import type { MemoryPrivateKeyStore } from '@veramo/key-manager';
-import { ICheqdCreateLinkedResourceResponse, ICheqdCreateLinkedResourceArgs, ResourcePayload, ISignInputs, IDIDManager, IKeyManager, ICredentialIssuer, ICredentialVerifier, IResolver, TAgent, IDataStore, ICheqd } from '@originvault/ov-types';
+import { ICheqdCreateLinkedResourceArgs, ResourcePayload, ISignInputs, IDIDManager, IKeyManager, ICredentialIssuer, ICredentialVerifier, IResolver, TAgent, IDataStore, ICheqd, IOVAgent } from '@originvault/ov-types';
 import fs from 'fs';
 import path from 'path';
 import { getDIDKeys } from './identityManager.js';
 import { CheqdDIDProvider } from '@cheqd/did-provider-cheqd';
+import { co2 } from "@tgwf/co2";
 
 let jsonFilePath;
+let dirId;
 
-const cleanUp = () => {
+const cleanUp = (collectionId?: string) => {
     if (jsonFilePath) fs.unlinkSync(jsonFilePath);
+    if(collectionId) {
+        fs.rmdirSync(collectionId, { recursive: true });
+    }
 }
 
 export async function createResource({ data, did, name, version, provider, agent, keyStore, resourceId, resourceType }: { data: any, did: string, name: string, version: string, provider: CheqdDIDProvider, agent: TAgent<IKeyManager & IDIDManager & ICredentialIssuer & ICredentialVerifier & IResolver & IDataStore & ICheqd>, keyStore: MemoryPrivateKeyStore, resourceId?: string, resourceType?: string, }) {    
     try {
-        const resolvedDid = await getDIDKeys(did);
+        const resolvedKeys = await getDIDKeys(did);
 
-        if (!resolvedDid) {
+        if (!resolvedKeys) {
             console.log("Could not resolve DID", did);
             return undefined;
         }
         
-        const { meta } = resolvedDid;
-        const { keyName: id, kid: key } = meta as { keyName: string, kid: string };
+        const { keyName: id, kid: key } = resolvedKeys;
         
-        const verificationMethod = await getVerifiedAuthentication(id);
+        const verificationMethod = await getVerifiedAuthentication(id as string);
 
         // Extract the last part of the DID string to use as the collectionId
-        const collectionId = id.split(':').pop() || '';
-       
+        const collectionId = (id as string).split(':').pop() || '';
+        dirId = collectionId;
         const fileRelativePath = uuidv5(name, uuidv5.URL); // Use sanitized name
         const resourceUUID = resourceId || uuidv5(fileRelativePath + new Date().toISOString(), uuidv5.URL);
-        const privateKey = await keyStore.getKey({ alias: key });
+        const privateKey = await keyStore.getKey({ alias: key as string });
 
         jsonFilePath = await generateResourceFile(collectionId, fileRelativePath, data);
         // Check if the file exists before reading
@@ -48,8 +52,6 @@ export async function createResource({ data, did, name, version, provider, agent
         }];
 
         const payload: ResourcePayload = {
-            did: id,
-            key: key,
             collectionId,
             id: resourceUUID,
             name,
@@ -61,7 +63,7 @@ export async function createResource({ data, did, name, version, provider, agent
             payload.version = version;
         }
 
-        const params: ICheqdCreateLinkedResourceArgs = {
+        const options: ICheqdCreateLinkedResourceArgs = {
             kms: 'local',
             network: CheqdNetwork.Mainnet,
             payload,
@@ -72,10 +74,19 @@ export async function createResource({ data, did, name, version, provider, agent
                 gas: '2000000',
             }
         }
+
+        const params = {
+            options,
+        }
+        
+        const co2Emission = new co2();
+        const co2EmissionResult = co2Emission.perByte(JSON.stringify(params).length, false);
+        
+        console.log(`🌱 ${name}@${version} - Resource size in carbon grams: ${co2EmissionResult.toFixed(5)}g`);
         
         try {
-            const result: ICheqdCreateLinkedResourceResponse = await provider.createResource({ options: params }, { agent });
-            cleanUp();
+            const result = await provider.createResource(params, { agent });
+            cleanUp(dirId);
             if (result) {
                 // Return the link to the cheqd resolver
                 return `https://resolver.cheqd.net/1.0/identifiers/${did}/resources/${resourceUUID}`;
@@ -83,13 +94,13 @@ export async function createResource({ data, did, name, version, provider, agent
             
             return undefined;
         } catch (error) {
-            cleanUp();
+            cleanUp(dirId);
             throw `Error creating resource: ${error}`;
         }
     }
     catch (error) {
         console.error("Check data formatting and RPC endpoint connection. Error creating resource:", error);
-        cleanUp();
+        cleanUp(dirId);
         return undefined;
     }
 }
@@ -110,7 +121,7 @@ export function generateResourceFile(dirPath: string, filePath: string, data: an
     return jsonFilePath;
 }
 
-export async function getResources({ did, agent }: { did: string, agent: TAgent<IKeyManager & IDIDManager & ICredentialPlugin & IResolver> }) {
+export async function getResources({ did, agent }: { did: string, agent: IOVAgent }) {
     const resolvedDid = await agent.resolveDid({ didUrl: did });
     console.log('DID', resolvedDid);
 }